test: add GPU and CPU viability tests for Markov and mask attacks
Adds empirical correctness tests to verify the Markov and mask rainbow
table attacks produce correct GPU chain output, plus CPU-only quality
and completeness checks.

New GPU test files:
- test_index_to_plaintext_markov.c/h + CL + Metal kernels: verifies
  index_to_plaintext_markov GPU kernel matches CPU reference for all
  indices of a synthetic 3-char model (13 test vectors)
- test_chain_markov.c/h: verifies crackalack_markov kernel endpoint
  matches CPU chain walk (5 test vectors)
- test_index_to_plaintext_mask.c/h + CL + Metal kernels: verifies mask
  i2p GPU path (is_mask=1) for ?d, ?d?d, ?l?d, ?u?l?d, ?a, ?d?d?d?d
- test_chain_mask.c/h: verifies crackalack kernel (is_mask=1) endpoint
  matches CPU chain walk for ?d?d, ?l?d, ?u?l?d masks

New CPU test groups:
- test_markov.c Group G (MQ-01 to MQ-04): Markov model quality -
  common-first ordering, monotone probability, top-N coverage, uniform
  corpus does not crash or produce duplicates
- test_mask.c Group H (MC-01 to MC-05): mask completeness - exhaustive
  enumeration of ?d?d, ?l?d, ?u?d keyspaces, boundary validity for
  ?l?d?u, and specifier disjointness for ?l/?u/?d/?s

All 22 test groups pass on Apple M3 Max (Metal backend).

+++ crackalack_rtc2rt.c
@@ -18,7 +18,6 @@
 #ifdef _WIN32
 #include <windows.h>
 #endif
-#include <inttypes.h>
 #include <stdio.h>
 
 #include "rtc_decompress.h"
@@ -27,7 +26,7 @@
 
 int main(int ac, char **av) {
   uint64_t *uncompressed_table = NULL;
-  uint64_t i = 0, num_chains = 0;
+  unsigned int i = 0, num_chains = 0;
   char *rtc_filename_input = NULL, *rt_filename_output = NULL;
   int ret = 0;
   FILE *f = NULL;
@@ -62,6 +61,6 @@ int main(int ac, char **av) {
 
   fclose(f);
 
-  printf("Successfully uncompressed %"PRIu64" chains in RTC file \"%s\" to RT file \"%s\".\n", num_chains, rtc_filename_input, rt_filename_output);
+  printf("Successfully uncompressed %u chains in RTC file \"%s\" to RT file \"%s\".\n", num_chains, rtc_filename_input, rt_filename_output);
   return 0;
 }
